Added code to fetch project id by request id
This logic was added in case it is not possible to pass the project id to this endpoint
from within the MarkDelivery context that this endpoint is being hooked into.

If we CAN pass the project id then we can remove the code which fetches a project summary
given a request id. If not then we can keep it as is.

Signed-off-by: Angelica Ochoa <15623749+ao508@users.noreply.github.com>