Remove dependency on cmo-metadb-shared-entities
LimsRest will no longer share models with the cmo-metadb project.

In an effort to keep the shared entities package minimal we did not want to project to
require any neo4j dependencies. After further progress in the cmo-metadb project
design and planning it was realized that we could not avoid requiring neo4j
dependencies in the shared entities package.

The models we had moved to the shared entities package have been refactored back
into the LimsRest project and organized in its own model package for ease of use.

Signed-off-by: Angelica Ochoa <15623749+ao508@users.noreply.github.com>